+++ ATM/telephone_mng/Telephone.cs
@@ -9,9 +9,9 @@ namespace ATM
     class Telephone
     {
         public string Number { get; private set; }
-        public double Balance { get; private set; }
+        public decimal Balance { get; private set; }
 
-        public Telephone(string telNumb,double balance)
+        public Telephone(string telNumb,decimal balance)
         {
             if (balance < 0)
                 throw new ArgumentException("Telephone balance cannot be less than 0");
@@ -22,7 +22,7 @@ namespace ATM
         }
 
         //adds money to telephone balance
-        public void AddToBalance(double money)
+        public void AddToBalance(decimal money)
         {
             if (money < 0)
                 throw new ArgumentException("Money to put on telephone balance can't be value less than 0");
